+++ ChangeLog.md
@@ -1,5 +1,6 @@
 2022-10
 
+- add `missionTemplates` and `providers` member resolution for `AssetClass` (graphql)
 - data sql scripts to initialize the database
   - `just update-sql-init-scripts` regenerates the ones that use entities from TethysDash:
     - Units of measure